BACKEND/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
openai>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.0
pytest>=7.0.0
gunicorn>=20.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyJWT>=2.8.0
twilio>=8.10.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
flask-socketio>=5.3.0
pytest-cov>=4.1.0